Add GCHeapMallocator
In preparation for the removal of GCAllocator.deallocate and
GCAllocator.reallocate, make that functionality available via a new
allocator with a different name.

Users of GCAllocator that wish to continue using manual @system memory
management instead of automatic @safe memory management can switch to
GCHeapMallocator.

For the rationale behind these changes, see PR #8554 and issue 23318.