Add `std.getopt.config.helpLike` to `std.getopt` module
The new option `std.getopt.config.helpLike` allows to use command-line arguments like
additional 'help' arguments.

The argument following this new directive is treated like the built-in '-h' or '--help'
arguments: Any previous `std.getopt.config.required` option is circumvented and the
`helpWanted` flag is set.

A common use-case in combination with a previously given `std.getopt.config.required`
option is an additional later commandline argument like `--version` to just output the
(git-)version of a programm and then return with no error. This is similiar to the
`--help` argument, which will also circumvent checks for `required`parameters and set the
`helpwanted`flag instead.

The option `std.getopt.config.required` is commonly used in commandline utilities and D scripts
to enforce required commandline arguments. It is also common to have extra arguments like
`--version`, which just output the the programm version.

This doesn't work right now, because of the `std.getopt.config.required` option for some
other argument. A workaround is to pass `--help`together with `--version`, which circumvents
the check for `std.getopt.config.required`.

For practical reasons it is desirable to just pass `--version`.