zlib: Add importC special case for va_start() on Win64
This fixes an undefined `__va_start` symbol for the zlib parts,
a problem on Win64 only so far.

The Microsoft headers use something like this (`vadefs.h`) for
x86_64:

```
void __cdecl __va_start(va_list* , ...);
```

The signature of their intrinsic isn't compatible with druntime's
`va_start` (which takes the first param as `out` ref, not as
explicit pointer), so I sadly don't see a way to handle this in
druntime's `__builtins.di` (or `importc.h`).