Update docs for hasSlicing with inifnite range
The requirement for opSlice on an infinite range to return the result of
take or takeExactly was removed in commit c828a08b64, which was included
in Phobos 2.067.0, released in March 2015. However, the documentation
was never updated to take this change into account.

This change updates the documentation and adds a unit test for the "new"
behavior.

Fixes bugzilla issue 24348.