Deprecate the range API functions on Nullable.
I don't know how acceptable it is to do this (the functionality was
added in 2.100.0 back in May 2022, so it's not particularly old, but
it's not exactly new either). However, I'm sick of dealing with bugs
related to Nullables being treated as ranges. We got rid of the alias
this on Nullable precisely to fix problems related to implict
conversions and Nullable, and making Nullable a range just added a whole
new set of them - ones which typically pop up in generic code, forcing
you to do stuff like check that a type is not an instance of a Nullable
or add an overload just for Nullables so that the range-based overload
doesn't take it.

Code that wants to get a range from Nullable can do so by slicing it,
just like you can with static arrays and other container types. So, code
that wants to continue to use Nullables as ranges can do so quite easily
by simply adding []. In the meantime, they'll get a deprecation message,
so no code will break - though unfortunately, the problems with Nullable
being a range will of course continue until we can actually remove the
deprecated functions down the road.