Make phobos tests compile with Nullable's range API deprecated.
Apparently, Phobos is built with -de, which then tends to break any
tests involving deprecated features (particularly with anything
involving type introspection). So, rather than getting deprecation
messages for the tests (or being able to slap deprecated on the unittest
block to get rid of them), the code fails to compile.

So, this updates the tests that should continue to exist but use [] to
get a range from a Nullable and removes the unittest block which is
specifically for testing that the range API on nullable works properly
(since there's no way to test it with -de being used).