Remove lib.sys.stdint.
We really don't want to be publicly importing a bunch of C types as part
of Phobos. It may make sense in some specific cases but not like this.
If anything, std.stdint should have been gone years ago, but I guess
that it got missed when we purged std.c, since it wasn't under std.c.