Add lib.sys.meta.
This adds lib.sys.meta, containing just AliasSeq and Alias for now.
They've been taken from std.meta more or less as-is, since they work
just fine as they are.

For Phobos v3, we really need a replacement for std.traits in order for
most stuff to work, so I was going to start there, adding symbols from
there piece by piece, adjusting them as necessary (or skipping them if
appropriate) so that we'd have at least the basic traits necessary for
everything else, but it turns out that some of the meta stuff is needed
as well for the unit tests and documentation of traits stuff, so I'm
starting by adding lib.sys.meta as a replacement for std.meta with the
two most basic items from std.meta.

So, this starts the process of building up lib.sys.meta and
lib.sys.traits piece by piece.