Add Map to lib.sys.meta.
This ports std.meta.staticMap to lib.sys.meta as Map.

The implementation is the same, but the documentation and tests have
been expanded.

As has been discussed before, the way that std.meta names things is
inconsistent, and the pattern that lib.sys.meta is going with is that
its templates will be CamelCased like types, which matches what some of
std.meta does but not all of it.