Add ConstOf, ImmutableOf, InoutOf, and SharedOf to lib.sys.traits.
These do the same as their std.traits counterparts, but I've beefed up
the documentation and examples a bit.

I'm not sure that it's worth adding SharedConstOf, SharedInoutOf, or
SharedConstInoutOf as well, since that seems kind of excessive to me -
though it looks like std.variant does use SharedConstOf in its tests -
but in either case, I think that those can be left for a later commit if
we want to keep them, since this is already adding four symbols.