Rename lib.sys.traits' Unqual to Unqualified.
The reason for this is because Unqual currently gets used heavily when
in many (most?) cases, Unconst should be used instead, because the code
in question is not supposed to be operating on shared, and thus
stripping off shared will either allow shared types in when they
shouldn't be (e.g. with a template constraint), or in the case of a
cast, it will actually strip off shared and the protections that it
provides, risking concurrency bugs. But a lot of D programmers are in
the habit of using Unqual rather than Unconst (probably in part because
Unqual has been around longer) when they simply want to remove const,
and by renaming it, we will hopefully discourage its use - or at least
make programmers think about it briefly rather than unthinkingly
grabbing Unqual.

Obviously, there are cases where Unqualified should still be used, but
hopefully, with the new name, it will be used less when it shouldn't be.

I've also adjusted the documentation on Unconst, Unqualified, and
Unshared to try to better clarify the situation with user-defined types
and member variables (since as Paul Backus has pointed out, it's often
incorrectly assumed that Unqual is guaranteed to strip const off of all
of the member variables when that's not actually true).