Rename lib to phobos.
As discussed in a recent planning meeting, we're renaming lib to phobos,
making it so that the root for Phobos v3 is phobos. So, that's what this
commit does.

However, on top of that, I removed the existing package.d files (both
lib/package.d and lib/sys/package.d) rather than fixing them, since I'm
inclined to think that we should be very selective about when we use
package.d, and I think that allowing something like import phobos; is a
mistake.

It wouldn't surprise me if we get screaming about that later, and maybe
we'll add it back at that point, but I'd rather not worry about
maintaining it as modules get added, and if we can get away with it, I
think that we should just not have it at all, since it's better practice
in general to limit the symbols that you import to something close to
what you're actually using, whereas import phobos; would be importing
everything. And having a package.d at each level like lib/sys/package.d
seems to be trying to do would complicate things even further.

We may very well end up with some package.d files in some cases, but I
think that that should be handled on a case-by-case basis - and that the
default should be to only have one module to import a symbol from.

So, with these changes, lib becomes phobos, and there are no longer any
package.d files in it.