Add isSignedInteger and isUnsignedInteger to phobos.sys.traits. (#8954)
They're similar to isSigned and isUnsigned from std.traits, but these
new traits are far more restrictive, because they're only true for the
four signed/unsigned built-in integer types, whereas isSigned and
isUnsigned accept a bunch of different types (including floating point
types and types that implicitly convert to real in the case of
isSigned).

From what I've looked at in Phobos v2, most uses of isSigned and
isUnsigned are used specifically when the code already knows that it's
operating on integer types anyway, so these traits should work for most
of those cases, and floating point types are always signed, so including
them in a check for signedness seems kind of redundant. I suppose that
isNumericSigned might make sense under some set of circumstances, since
it would allow for a single template instantiation instead of two like
you'd get with isSignedInteger!T || isFloatingPoint!T, but at least for
now, I think that isSignedInteger and isUnsignedInteger are enough.