Add isImplicitlyConvertible and isQualifierConvertible to phobos v3. (#8955)
Both traits do exactly the same thing as their std.traits counterparts,
though these have far more documentation and examples.

I'm not a big fan of isImplicitlyConvertible, since it's usually a bad
idea to test for implicit conversions in template constraints, and the
is expression is incredibly simple to boot, but a trait _is_ needed when
a template predicate is needed (e.g. for the templates in
phobos.sys.meta), and having isImplicitlyConvertible allows for a good
place to put documentation on implicit conversions in general (as well
as to warn against testing for them in template constraints).

isQualifierConvertible isn't actually used much in Phobos v2 right now
(in most cases right now, Unqual is used with an is expression instead),
but it is used by the range traits, and there probably are a number of
cases where it would be better to use it than Unqualified or Unconst
with an is expression. But it's new enough that I don't think that many
of us have really thought through how best to use it yet. Either way, if
we want to be able to do something like isInputRange!(R, const char),
then it's a good trait to have.