Adds Instantiate, ApplyLeft, and ApplyRight to phobos.sys.meta.
This also makes it so that isImplicitlyConvertible and
isQualifierConvertible in phobos.sys.traits can be partially
instantiated, since I missed that before.

Instantiate is basically identical to the one in std.meta
implementation-wise, but the documentation and unit tests have been
improved.

ApplyLeft and ApplyRight should be functionally the same as their
std.meta counterparts, but they have a simpler implementation than their
std.meta counterparts (I'm pretty sure that the extra complexity in the
old implementation is to work around built-in types not working with
aliases, which was fixed a while ago). Of course, their documentation
and tests have also been improved.