Add And, Not, and Or to phobos.sys.meta.
These correspond to templateAnd, templateNot, and templateOr from
std.meta. The names have been shortened and made PascalCase to make them
consistent with our normal naming rules. It's PascalCase rather than
camelCase, because the result is a template rather than a value, and we
don't normally prefix symbols with "template" just because they're
templates.

The implementation of Not is unchanged. However, unlike their std.meta
counterparts, And and Or do not short-circuit their evaluation. This makes
them consistent with the rest of phobos.sys.meta (whereas std.meta is
inconsistent about whether its templates short-circuit their evaluation),
but it's also because they're implemented to be iterative rather than
recursive, which should make them more efficient.