Fix Bugzilla issue 24704: Improve error messages for from*String in std.datetime.
The documentation for the from*String functions specifies what they
accept, and the fromISO*String functions accept ISO or ISO extended
strings, but it is technically the case that there are ISO and ISO
extended strings which they do not accept (since the standard specifies
what the fields should look like for each string but doesn't specify
that all fields must be present or that any code parsing them must
accept all variants).

So, technically, the error messages which have said that the given
strings are not valid ISO or ISO extended strings are not necessarily
correct. So, this changes the error messages to remove that ambiguity.