std.process: Do not kill process group 2 in unittest
When passing a Pid object to tryWait(), its processId will be changed to
Pid.terminated, which happens to be -2. Passing this value to kill()
will kill the process group 2 on Posix systems. If you run the build and
test suite using this process group, the test case will kill the entire
build.

Co-authored-by: Emanuele Rocca <ema@debian.org>
Co-authored-by: Jochen Sprickerhof <git@jochen.sprickerhof.de>
Bug-Debian: https://bugs.debian.org/1089007