Add phobos.sys.traits.hasIndirections. (#10634)
It does the same thing as the std.traits version. They both alias the
one in druntime. However, the documentaton and tests have been beefed up
a bit.

I also added assertWithQualifiers as a private test helper, which it may
make sense to add later as a public test helper somewhere, but we don't
currently have any module or package for that sort of thing, so for now,
it's private. Some of the existing tests should probably also be
refactored to use it, but I'm not worrying about that as part of these
changes.